Inline inMemSnapshotStore
